microbitv2: Refactory board id and Kl27 timer code into their own files.
This should result in no change in behaviour.
