+++ source/board/microbitv2/kl27z/pwr_mon.c
@@ -20,6 +20,8 @@
  */
 
 #include "pwr_mon.h"
+#include "timer_wait.h"
+
 #include "adc.h"
 #include "fsl_adc16.h"
 #include "fsl_pmc.h"
@@ -39,10 +41,6 @@ static void pwr_mon_bandgap_init(void);
 static uint32_t pwr_mon_read_vbg(uint32_t channelGroup);
 static uint32_t pwr_mon_adc_to_mv(uint32_t raw_adc);
 
-void TPM0_IRQHandler(void);
-
-extern volatile bool tpmIsrFlag;
-extern uint32_t tpm_source_clock;
 
 void pwr_mon_init(void)
 {
@@ -50,9 +48,10 @@ void pwr_mon_init(void)
         .pinDirection = kGPIO_DigitalOutput,
         .outputLogic = 0U
     };
-    
+
     adc_init();
-    
+    timer_wait_init();
+
     // Configure VMON_BAT and RUN_VBAT_SENSE
     GPIO_PinInit(PIN_RUN_VBAT_SENSE_GPIO, PIN_RUN_VBAT_SENSE_BIT, &pin_config);
     PORT_SetPinMux(PIN_RUN_VBAT_SENSE_PORT, PIN_RUN_VBAT_SENSE_BIT, kPORT_MuxAsGpio);
@@ -90,23 +89,22 @@ power_source_t pwr_mon_get_power_source(void) {
     return power_source;
 }
 
-uint32_t pwr_mon_get_vbat_mv(void) {
+uint32_t pwr_mon_get_vbat_mv(void)
+{
     /* Set timer period 3ms*/
-    TPM_SetTimerPeriod(TPM0, USEC_TO_COUNT(3000U, tpm_source_clock));
+    timer_wait_set_period(3000);
     // Enable voltage divider to take measurement
     GPIO_PinWrite(PIN_RUN_VBAT_SENSE_GPIO, PIN_RUN_VBAT_SENSE_BIT, 1);
     // Add a ~3ms delay to allow the 100nF capacitors to charge to about 3*RC.
-    TPM_StartTimer(TPM0, kTPM_SystemClock);
-    while (false == tpmIsrFlag);
-    tpmIsrFlag = false;
+    timer_wait();
 
     uint32_t bat_adc = adc_read_channel(0, PIN_VMON_BAT_ADC_CH, PIN_VMON_BAT_ADC_MUX);
     // Disable voltage divider
     GPIO_PinWrite(PIN_RUN_VBAT_SENSE_GPIO, PIN_RUN_VBAT_SENSE_BIT, 0);
-    
+
     // Compensate for voltage divider with ratio of 11
     bat_adc = bat_adc * 11;
-    
+
     return pwr_mon_adc_to_mv(bat_adc);
 }
