+++ source/board/microbitv2/kl27z/timer_wait.h
@@ -1,0 +1,33 @@
+/**
+ * @file    timer_wait.h
+ * @brief
+ *
+ * DAPLink Interface Firmware
+ * Copyright 2021 Micro:bit Educational Foundation
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef TIMER_WAIT_H_
+#define TIMER_WAIT_H_
+
+#include <stdint.h>
+
+
+void timer_wait_init(void);
+void timer_wait_set_period(uint32_t microsec);
+void timer_wait(void);
+
+
+#endif /* TIMER_WAIT_H_ */
